Avoid error-prone mixing of signed and unsigned
Resolved all warnings that '-Wsign-conversion' would elicit in a 32-bit
compile with i686-w64-mingw32-gcc-7.3, except the very large number that
  grep 'error:' |sed -e '/size_type/d'
filters out (notably due to the absence of 'vector::operator[](int)').

* fenv_lmi_x86.hpp: Incidentally used std::bitset::operator[] uniformly,
  removing a workaround for an ancient compiler.
* ihs_acctval.cpp: Replaced assertion + operator[] with at().
* mc_enum_metadata.hpp: Removed a puzzling comment. Rewrote a static
  assertion to clarify its actual meaning.
* miscellany.cpp: In count_newlines(), it is not necessarily ideal to
  return the result of std::count() as an int, but returning std::size_t
  certainly involved an implicit sign conversion. (The same may be said
  of returning std::ptrdiff_t as int in several other files.)
* test_coding_rules.cpp: Incidentally used an initializer-list for a
  std::set; the same technique (N1509) should be used elsewhere.
* timer.cpp: Preferred 'int' for lmi's own timing function, casting it
  as needed for OS calls. Sleeping for more than 2^31 seconds is not a
  realistic use case; of course, bourn_cast<>() throws an exception if
  anyone tries to do that.
* vector_test.cpp: Removed casts that, although correct, were curiously
  used only on the second and third of three similar successive lines.

+++ vector_test.cpp
@@ -30,7 +30,6 @@
 #include "timer.hpp"
 
 #include <algorithm>
-#include <cstddef>                      // size_t
 #include <functional>
 #include <iomanip>
 #include <ios>
@@ -287,8 +286,8 @@ namespace
     simple_array0 g_w(g_array_length);
 
     std::valarray<double> g_va_u(g_array_length);
-    std::valarray<double> g_va_v(static_cast<std::size_t>(g_array_length));
-    std::valarray<double> g_va_w(static_cast<std::size_t>(g_array_length));
+    std::valarray<double> g_va_v(g_array_length);
+    std::valarray<double> g_va_w(g_array_length);
 } // Unnamed namespace.
 
 void mete_c()
